requirements.txt
google-generativeai
fastapi
qdrant-client
psycopg
uvicorn
python-dotenv
pytest